requirements.txt
asgiref==3.5.0
autopep8==1.6.0
bcrypt==3.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
colorama==0.4.4
confusable-homoglyphs==3.2.0
crispy-bootstrap5==0.6
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
dj-rest-auth==2.2.3
Django==4.0.2
django-allauth==0.48.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-heroku==0.3.1
django-ipware==4.0.2
django-model-utils==4.2.0
django-on-heroku==1.1.2
django-registration==3.2
django-rename-app==0.1.3
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.13.1
docopt==0.6.2
docutils==0.18.1
drf-extra-fields==3.2.1
Fabric3==1.14.post1
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.2
Jinja2==3.0.3
keyring==23.5.0
MarkupSafe==2.1.0
oauthlib==3.2.0
packaging==21.3
paramiko==2.9.2
Pillow==9.0.1
pipenv==2022.1.8
pipreqs==0.4.11
pkginfo==1.8.2
platformdirs==2.5.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.7
python-decouple==3.6
python-dotenv==0.19.2
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2021.3
pywin32-ctypes==0.2.0
readme-renderer==32.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.63.0
twine==3.8.0
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.2
virtualenv-clone==0.5.7
webencodings==0.5.1
whitenoise==6.0.0
yarg==0.1.9
zipp==3.7.0
